+++ samples/sampler/stories/default/Scroller.js
@@ -14,9 +14,9 @@ import css from './Scroller.module.less';
 const prop = {
 	direction: ['both', 'horizontal', 'vertical'],
 	focusableScrollbarOption: {
-		'true': true,
-		'false': false,
-		'byEnter': 'byEnter'
+		false: false,
+		true: true,
+		'&quot;byEnter&quot;': 'byEnter'
 	},
 	scrollbarOption: ['auto', 'hidden', 'visible'],
 	scrollModeOption: ['native', 'translate']
@@ -30,7 +30,7 @@ storiesOf('Sandstone', module)
 		() => {
 			const
 				direction = select('direction', prop.direction, ScrollerConfig),
-				focusableScrollbar = select('focusableScrollbar', prop.focusableScrollbarOption, ScrollerConfig),
+				focusableScrollbar = prop.focusableScrollbarOption[select('focusableScrollbar', ['false', 'true', '"byEnter"'], ScrollerConfig)],
 				horizontalScrollbar = select('horizontalScrollbar', prop.scrollbarOption, ScrollerConfig),
 				verticalScrollbar = select('verticalScrollbar', prop.scrollbarOption, ScrollerConfig);
 			return (