+++ CHANGELOG.md
@@ -4,6 +4,51 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Fixed
+
+- `sandstone/Input` to support accessibility features
+- `sandstone/Panels` to properly restore focus after a transition
+
+## [1.0.0-beta.7] - 2020-06-16
+
+### Added
+
+- `sandstone/Dropdown` prop `title` to optionally display a heading above the component
+- `sandstone/FixedPopupPanels` and `sandstone/FlexiblePopupPanels` prop `fullHeight` to force these components to always stretch to the screen edges
+- `sandstone/Icon` prop `flip` value `"auto"` to automatically flip the icon horizontally for RTL locales
+- `sandstone/TooltipDecorator` prop `tooltipType` to support new transparent label-style tooltips
+- `sandstone/TooltipDecorator` prop `tooltipMarquee` to support marquee
+
+### Changed
+
+- `sandstone/Dropdown` prop `title` to `placeholder` to display a value within the component when no selection has been made
+- `sandstone/Input` to highlight activated number cells
+- `sandstone/Panel` and `sandstone/WizardPanels` support for reference forwarding to obtain a reference to each component's root node
+
+### Fixed
+
+- `sandstone/TimePicker` to read out 'hour' or 'minute' when it is focused or its value is changed
+- `sandstone/DatePicker` to read out 'day', 'month', or 'year' when it is focused or its value is changed
+- `sandstone/Alert` to center its content when `type="fullscreen"`
+- `sandstone/Button` flashing when switching `selected` on and off
+- `sandstone/CheckboxItem`, `sandstone/FormCheckboxItem`, `sandstone/RadioItem`, and `sandstone/SwitchItem` slots margins
+- `sandstone/ContextualMenuDecorator` to not be read as an alert when rendered
+- `sandstone/ContextualPopupDecorator` to position itself correctly when `direction` is changed
+- `sandstone/DayPicker` format for locales that do not start the week on Sunday
+- `sandstone/Dropdown` to properly read the focused item
+- `sandstone/FixedPopupPanels` layout in RTL locales
+- `sandstone/FixedPopupPanels` to support accessibility properly
+- `sandstone/FixedPopupPanels` to flex to the content size and invoke scrolling (when using `sandstone/Scroller`) when the content is too big
+- `sandstone/Input` to update `invalidTooltip` to the latest design
+- `sandstone/Panel` and `sandstone/WizardPanels` to not read out the Panel title after closing a dropdown
+- `sandstone/TabLayout` to restore focus to the selected tab when expanding without icons
+- `sandstone/TabLayout` performance when focusing items in the layout
+- `sandstone/ThemeDecorator.AccessibilityDecorator` not to overwrite its `skinVariants` prop
+- `sandstone/VirtualList` focus when 5way directional keys are quickly and consecutively pressed
+- `sandstone/WizardPanels` to use `sandstone/Skinnable`
+
+## [1.0.0-beta.6] - 2020-06-08
+
 ### Removed
 
 - `sandstone/Alert` prop `subtitle`
@@ -11,11 +56,25 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Added
 
 - `sandstone/KeyGuide` support for color keys
+- `sandstone/Scroller` props `horizontalScrollThumbAriaLabel` and `verticalScrollThumbAriaLabel` to provide customization of the hint string read when a scroll thumb is focused
+
+### Changed
+
+- `sandstone/Icon` supported icon list, adding new icons and removing unused ones
 
 ### Fixed
 
 - `sandstone/Alert` to use multi-line content when `"fullscreen"`
+- `sandstone/Checkbox`, `sandstone/CheckboxItem`, `sandstone/Switch`, `sandstone/SwitchItem`, `sandstone/RadioItem` and `sandstone/FormCheckboxItem` to read out as selected value
+- `sandstone/Dropdown` to retain correct focus when `selected` or `children` change
+- `sandstone/Dropdown` to show an item fully when the item gets focus
 - `sandstone/FixedPopupPanels` and `sandstone/FlexiblePopupPanels` to properly respond to back button presses
+- `sandstone/Panels.Header` style to match latest designs
+- `sandstone/Popup` to always remove the scrim when closed
+- `sandstone/Scroller` and `sandstone/VirtualList` to show the scrollbar initially
+- `sandstone/Scroller` and `sandstone/VirtualList` to show the horizontal overscroll effect properly in RTL locales
+- `sandstone/TabLayout` button sizes to match the latest designs
+- `sandstone/WizardPanels` to respect using `spotlight/SpotlightContainerDecorator.spotlightDefaultClass` to determine the default focus
 
 ## [1.0.0-beta.5] - 2020-06-01
 
@@ -34,7 +93,6 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Changed
 
 - `sandstone/FlexiblePopupPanels` to provide a close button on the first panel and navigational buttons on each panel
-- `sandstone/Icon` supported icon list, adding new icons and removing unused ones
 
 ### Fixed
 
@@ -81,7 +139,6 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/Tooltip` style to match latest designs
 - `sandstone/VideoPlayer` to jump back when using the 5-way left key
 - `sandstone/VirtualList` to support navigation with spottable children inside an item
-- `sandstone/Checkbox`, `sandstone/CheckboxItem`, `sandstone/Switch`, `sandstone/SwitchItem`, `sandstone/RadioItem` and `sandstone/FormCheckboxItem` to read out as selected value
 
 ## [1.0.0-beta.3] - 2020-05-11